+++ src/Simplify.Web/Responses/Json.cs
@@ -1,11 +1,10 @@
 ﻿using System.Text.Json;
-using Simplify.Web;
+using System.Threading.Tasks;
 
-namespace SampleApp.Angular.Responses;
+namespace Simplify.Web.Responses;
 
 /// <summary>
-/// Provides controller JSON response (send only JSON string to response)
-/// This is sample Json response, it is recommended to use https://github.com/SimplifyNet/Simplify.Web.Json package instead
+/// Provides controller JSON response using System.Text.Json (send only JSON string to response)
 /// </summary>
 /// <remarks>
 /// Initializes a new instance of the <see cref="Json"/> class.
@@ -14,6 +13,15 @@ namespace SampleApp.Angular.Responses;
 /// <param name="statusCode">The HTTP response status code.</param>
 public class Json(object objectToConvert, int statusCode = 200) : ControllerResponse
 {
+
+	/// <summary>
+	/// Gets or sets the default JsonSerializerOptions
+	/// </summary>
+	/// <value>
+	/// The default JsonSerializerOptions.
+	/// </value>
+	public static JsonSerializerOptions? DefaultOptions { get; set; } = null;
+
 	/// <summary>
 	/// Gets the HTTP response status code.
 	/// </summary>
@@ -25,18 +33,12 @@ public class Json(object objectToConvert, int statusCode = 200) : ControllerResp
 	/// <summary>
 	/// Processes this response
 	/// </summary>
-	/// <returns></returns>
 	public override async Task<ResponseBehavior> ExecuteAsync()
 	{
 		Context.Response.ContentType = "application/json";
 		Context.Response.StatusCode = _statusCode;
 
-		await ResponseWriter.WriteAsync(Context.Response,
-			JsonSerializer.Serialize(objectToConvert,
-				new JsonSerializerOptions
-				{
-					PropertyNamingPolicy = JsonNamingPolicy.CamelCase
-				}));
+		await ResponseWriter.WriteAsync(Context.Response, JsonSerializer.Serialize(objectToConvert, DefaultOptions));
 
 		return ResponseBehavior.RawOutput;
 	}